Fix reading array of matrices.
We need to specify the shape of the desired matrices. Otherwise
the routine just returns the full 1D array for the first node
it reads.

The method is a bit clumy (we test against all primitive types)
and works with a copy, but this will be optimized and cleaned
later (e.g. using mastodon-collections to wrap a list of
double[] around a single array of generic primitives). Later.